​

+++ File.cs
@@ -6,33 +6,25 @@ namespace Namespace
 {
     public static class Class
     {
-        private static System.UInt64 Number = 0;
-        private static System.UInt64 Number2 = 0;
-        private static System.String File = "File.txt";
-        private static System.String File2 = "File.txt";
-        private static System.String Contents = System.String.Empty;
-        private static System.Exception Exception;
-        private static readonly System.UInt64 Number3 = System.UInt64.MaxValue;
-        private static void Main(System.String[] _)
+        static System.UInt64 Number = 0;
+        static System.UInt64 Number2 = 0;
+        static System.String File = "File.txt";
+        static System.String File2 = "File.txt";
+        static System.String Contents = System.String.Empty;
+        static System.Exception Exception;
+        static readonly System.UInt64 Number3 = System.UInt64.MaxValue;
+        static void Main()
         {
-            if (!System.IO.File.Exists(Namespace.Class.File))
-            {
-                System.IO.File.WriteAllText(Namespace.Class.File, "Goodbye Cruel World!" + System.Environment.NewLine);
-            }
+            if (!System.IO.File.Exists(Namespace.Class.File)) System.IO.File.WriteAllText(Namespace.Class.File, "Goodbye Cruel World!" + System.Environment.NewLine);
             System.Console.Out.WriteLine("WARNING! THIS A FILE CLONER! " +
                   "EXECUTING THIS WILL RESULT IN A FILE OVERLOAD!");
+            System.Console.Out.Flush();
             System.Console.Out.WriteLine("ARE YOU SURE YOU WANT TO CONTINUE? Y OR N");
-            if (System.Console.ReadLine().ToUpper().Contains("Y"))
-            {
-                Namespace.Class.End();
-            }
-            else
-            {
-                System.Environment.Exit(0);
-                return;
-            }
+            System.Console.Out.Flush();
+            if (System.Console.ReadLine().ToUpper().Contains("Y")) Namespace.Class.End();
+            else System.Environment.Exit(0);
         }
-        private static void End()
+        static void End()
         {
             System.Console.Clear();
             Namespace.Class.Contents = System.IO.File.ReadAllText(Namespace.Class.File);
@@ -43,6 +35,7 @@ namespace Namespace
                 {
                     System.IO.File.AppendAllText(Namespace.Class.File, Namespace.Class.Contents);
                     System.Console.Out.Write('\r' + (Namespace.Class.Number2 + 1).ToString());
+                    System.Console.Out.Flush();
                     System.Console.Title = (Namespace.Class.Number2 + 1).ToString();
                     System.Console.Out.Flush();
                     try
@@ -55,6 +48,7 @@ namespace Namespace
                     {
                         Namespace.Class.Exception = ex;
                         System.Console.Out.WriteLine("Exception occurred: {0}", Namespace.Class.Exception);
+                        System.Console.Out.Flush();
                         System.IO.File.Delete(Namespace.Class.File2);
                         Namespace.Class.End();
                     }
@@ -63,11 +57,13 @@ namespace Namespace
                 {
                     Namespace.Class.Exception = ex;
                     System.Console.Out.WriteLine("Exception occurred: {0}", Namespace.Class.Exception);
+                    System.Console.Out.Flush();
                     System.IO.File.Delete(Namespace.Class.File);
                     Namespace.Class.End();
                 }
             }
             System.Console.Out.WriteLine();
+            System.Console.Out.Flush();
         }
     }
 }
